requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
python-dotenv==0.19.0
PyPDF2==1.26.0
pdfminer.six==20201018
pandas==1.5.3
numpy==1.23.5
sqlalchemy==1.4.22
WTForms==2.3.3
requests==2.26.0
